feat: add support for debugging Titanium applications (#116)
* feat(debugger): declare debugger, hook up into vscode

To debug the debugger (_inception noise_), run the Extension + Server launch config, and ensure debugServer: 4711 is set in your launch config for the titanium app

* build: add node internals to skipFiles for debugging

* feat(debugger/android): working version of classic app debugging

* feat(debugger/android): prompt for device if none provided

* feat(debugger): terminate build on debug cancel

* feat(debugger): handle error exit of the cli

* feat(debugger/android): working version of classic app debugging

* feat(debugger): handle remapping of sourcemaps for alloy

This requires a change to add the sourceRoot property to the alloy source maps

* fix(debugger): hant statdle when no device was selected

* refactor(debugger): correct bad import

* fix: handle if channel is not created yet

refactor: remove redundant code

* fix(debugger): add skip js minify flag when debugging

* feat: add support for debugging ios simulators and devices

* feat: check for ios_webkit_debug_proxy prior to starting debug session

* fix(debugger/ios): do not prompt for certificate info on attach request

* refactor: cleanup initialConfiguration and configurationSnippets

* refactor(debugger): move to using a DebugConfigurationProvider

This allows us to resolve the extra info we need prior to starting the debug flow and avoiding the awkward communication processes we have

* refactor(debugger): drop Debugger from environment label

* fix(debugger): uri encode app name when transforming the source url

The source url we're provided is URI encoded so we need to make sure we match things like spaces

* docs(debugger): start documenting how to use the debugger

* fix(debugger): require CLI 7.1.0 or greater for debugging

* docs: add some gifs to docs, improve install steps

* feat(debugger): add last debug state to target selection

* feat(debugger): try a best attempt at validating port is free

* refactor: move iwdb check to debug provider, add docs link

* doc: fix typo

* fix(debugger/config): correct iwdb command name

Im dumb

* fix(debugger/android): try to kill adb port when finishing debugging

While this wont work on emulators as we dont have the right device id for use in adb, adb seems pretty good at reusing the port if we choose it again

* refactor(debugger): improve the port checking to be more robust

* docs(debugging): add note about enabling web inspector on ios device

* fix(debugger/ios): disconnect when error occurs during attach

* fix(debugger/ios): increase retries on polling

When an iOS sim is booted from cold (i.e. it hasn't been booted for a while), we've seen timeouts where the app JSContext doesn't show in iwdb in time, an increase in the retries seems to be giving everything the correct amount of time to occur and get setup

* fix(android/debugger): ensure contextId is present on evaluate call

It appears Android does not correctly handle no contextId being provided (the property is optional on the request) so provide a contextId of 1. I need to file a ticket to have this fixed in the inspector client on the SDK side

* fix(debugger/android): remove forwarded port on android emulator

This prevents an issue when debugging on android then switching to debugging on ios without killing the emulator, leaving the forwarded port open which would cause the iOS connection to just never succeed as it's trying to connect to the android emulator

* docs(debugging): add note about debug repl, update images

* chore: update url to docs

Will be live once PR is merged