refactor(debugger)!: use task for building
This makes the debugging use a preBuildTask to build the application, this task has a problem matcher setup to look for the application launch message which will then trigger the connection of the debugger. All of the build prompting and configuration is now done by the task. Configuring the launch task target and device id has been removed in favour of using a custom task to define these

BREAKING CHANGE: Setting deviceId and target in a launch configurations is no longer supported. Use the preLaunchTask to set targets and device details or use the prompting flow

	modified:   src/common/extensionProtocol.ts