feat(commands): support multi-root workspaces
Commands (and features backed by commands) will now prompt for the which folder to run in when the
folder that the action should be ran in is not known (e.g there are multiple Titanium projects in
the workspace). Features that support multiple project types (e.g building and packaging) will now
support both app and module projects open in the same workspace.

This does not move the intellisense handling and completions generation to be mult-root workspace
aware, that will be done in a follow on commit as it is a different scope to this.