feat: add support for vscode workspace trust configuration
This add support for the upcoming workspace trust configuration, the support is rather blunt in
that the extension will not work unless the workspace has been trusted. This was chosen because
of the variety of ways this extensions works i.e. installing tooling, building project (which can
also have hooks that run code), so it made more sense to stick with the stricter and more secure
option

BREAKING CHANGE: Raises the minimum VS Code version to 1.56.0