feat: add keystore creation and keystore information storage
This adds support for creating a keystore and storing the keystore information. When packaging an
Android application there is now an extra "Create keystore" option. This will trigger a UI where
the information can be input, and optionally saved.

Also, when packaging for Android an extra "store information" dialog will be shown, this will store
the keystore information, using the path as they key, using the secure storage for the device.

Closes #89